Changed namespace and added an extra layer of security.

+++ w00p.php
@@ -8,19 +8,21 @@
     Author URI:   https://borgenfalk.se
     License:      MIT
     License URI:  https://github.com/tommyskott/please-do-track/blob/master/LICENSE
-    Text Domain:  BS_CPT
+    Text Domain:  BS_PLS
     Domain Path:  /languages
   */
 
   defined('ABSPATH') or die('No script kiddies please!');
-  define('BS_CPT_TXT_DOMAIN', 'BS_CPT');
+  define('BS_PLS_TXT_DOMAIN', 'BS_PLS');
 
   register_activation_hook(__FILE__, function(){
-    // clear oembed caches
-    global $wpdb;
-    $wpdb->query( 'DELETE FROM `' . $wpdb->postmeta . '` WHERE `meta_key` LIKE "_oembed%"' );
+    if(is_admin()){
+      // clear oembed caches
+      global $wpdb;
+      $wpdb->query( 'DELETE FROM `' . $wpdb->postmeta . '` WHERE `meta_key` LIKE "_oembed%"' );
+    }
 
-    BS_CPT_init();
+    BS_PLS_init();
     flush_rewrite_rules();
   });
 
@@ -29,19 +31,19 @@
   });
 
   add_action('init', function(){
-    BS_CPT_init();
+    BS_PLS_init();
   });
 
-  function BS_CPT_init(){
-    load_plugin_textdomain(BS_CPT_TXT_DOMAIN, false, basename(dirname(__FILE__)) . '/languages');
+  function BS_PLS_init(){
+    load_plugin_textdomain(BS_PLS_TXT_DOMAIN, false, basename(dirname(__FILE__)) . '/languages');
   }
 
-  function BS_CPT_oembed($provider){
+  function BS_PLS_oembed($provider){
     if(strpos($provider, 'vimeo.com') !== false){
       $provider = remove_query_arg('dnt', $provider);
     }
     return $provider;
   }
-  add_filter('oembed_fetch_url', 'BS_CPT_oembed');
+  add_filter('oembed_fetch_url', 'BS_PLS_oembed');
 
 ?>
